django-app/aOHS/requirements.txt
django~=3.2.9
channels~=3.0.4
channels-redis
asgiref~=3.4.1
opencv-python~=4.5.4.60
django-bootstrap-v5
python-ffmpeg-video-streaming>=0.1
django-cors-headers
Pillow~=9.0.0
requests~=2.26.0